* Increase i2c rx buffer to 64 bytes.
I have always used cyz_rgb with the Arduino. Unfortunately it seems the Wire library which is the most commonly used method of communicating with i2c devices on the arduino platform is somewhat weak. It defaults to 100khz bus speed and still has frequent trouble, weak error handling, and no support for repeat start.

Since we have a bit of available ram, I am increasing the rx buffer for the hardware twi bus in preparation for a new cyz_rgb arduino library that will not use Wire and run at th full 400khz. The larger buffer should help with operation at the higher speed and improve reliability when a large number of devices are on the i2c network.